Refactor filters into separate widgets and update README
