+++ pubspec.yaml
@@ -75,7 +75,6 @@ flutter:
   uses-material-design: true
   assets:
     - assets/images/
-    - assets/icons/
 
   # To add assets to your application, add an assets section, like this:
   # assets:
